requirements.txt
ansicon==1.89.0
blessed==1.21.0
customtkinter==5.2.2
darkdetect==0.8.0
easyocr==1.7.2
editor==1.6.6
filelock==3.18.0
fsspec==2025.5.1
imageio==2.37.0
inquirer==3.4.0
Jinja2==3.1.6
jinxed==1.3.0
keyboard==0.13.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-bidi==0.6.6
pytweening==1.2.0
PyYAML==6.0.2
readchar==4.2.1
rich==14.0.0
runs==1.2.2
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
sympy==1.14.0
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.0
wcwidth==0.2.13
xmod==1.8.1
